python/requirements.txt
aiohttp==3.7.4.post0
aioitertools==0.8.0
async-timeout==3.0.1
attrs==21.2.0
boto3
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
idna==3.2
jmespath==0.10.0
kafka-python==2.0.2
kubernetes==18.20.0
multidict==5.1.0
pycparser==2.20
PyYAML==5.4.1
ruamel.yaml==0.16.12
s3fs==2021.8.0
s3transfer
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
yarl==1.6.3
